Rename TM trafficopswrapper to towrap
