+++ traffic_monitor_golang/traffic_monitor/todata/todata.go
@@ -23,7 +23,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"github.com/apache/incubator-trafficcontrol/lib/go-tc"
-	towrap "github.com/apache/incubator-trafficcontrol/traffic_monitor_golang/traffic_monitor/trafficopswrapper"
+	"github.com/apache/incubator-trafficcontrol/traffic_monitor_golang/traffic_monitor/towrap"
 	"regexp"
 	"strings"
 	"sync"
