+++ traffic_monitor_golang/traffic_monitor/towrap/towrap.go
@@ -1,4 +1,4 @@
-package trafficopswrapper
+package towrap
 
 /*
  * Licensed to the Apache Software Foundation (ASF) under one
